requirements.txt
# Core Dependencies
python>=3.11

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
streamlit-autorefresh>=0.0.6

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0
fastparquet>=2023.10.0

# NLP & ML
transformers>=4.35.0
torch>=2.1.0
sentencepiece>=0.1.99

# Data Collection
feedparser>=6.0.10
requests>=2.31.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
scipy>=1.11.0

# Configuration
pyyaml>=6.0.1

# Database (Optional)
sqlalchemy>=2.0.23

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Storage (Optional)
boto3>=1.28.0  # S3 지원

# Development & Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0

